gcp_fns/image_transform/requirements.txt
opencv-python==4.2.0.34
numpy==1.18.2
google-cloud-storage==1.27.0